+++ R/test_connection.R
@@ -8,10 +8,13 @@ test_connection <- function(master = master,
   state$open <- TRUE
 
   structure(class = c("spark_connection", "test_connection"), list(
+    # spark_connection
     master = master,
     config = config,
     state = state,
-    spark_context = test_jobj_create(list())
+    spark_context = test_jobj_create(list()),
+    state = new.env()
+    # test_connection
   ))
 }
 
@@ -90,4 +93,20 @@ sdf_copy_to.test_connection <- function(sc, x, ...) {
 #' @export
 create_hive_context.test_connection <- function(sc) {
   test_jobj_create(sc)
+}
+
+#' @export
+hive_context.test_connection <- function(sc) {
+  if (is.null(sc$hive_context))
+    sc$hive_context <- create_hive_context(sc)
+
+  sc$hive_context
+}
+
+#' @export
+spark_session.test_connection <- function(sc) {
+  if (is.null(sc$hive_context))
+    sc$hive_context <- create_hive_context(sc)
+
+  sc$hive_context
 }